refactor: Add package path on the remote offset manager
- SpringBootApplication requires package.
