+++ offset-manager-remote/build.gradle
@@ -23,6 +23,7 @@ dependencies {
   annotationProcessor("org.projectlombok:lombok:1.18.38")
   testCompileOnly("org.projectlombok:lombok:1.18.38")
   testImplementation platform('org.junit:junit-bom:5.10.0')
+  testImplementation 'org.springframework.boot:spring-boot-starter-test'
   testImplementation 'org.junit.jupiter:junit-jupiter'
   testImplementation 'org.assertj:assertj-core:3.27.6'
   testImplementation 'org.mockito:mockito-core:5.20.0'
