+++ offset-manager-remote/src/main/java/localoffsetmanager/RemoteOffsetManagerApplication.java
@@ -1,0 +1,8 @@
+package localoffsetmanager;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RemoteOffsetManagerApplication {
+
+}
