+++ offset-manager-remote/src/main/java/localoffsetmanager/config/TopicConfig.java
@@ -1,4 +1,4 @@
-package config;
+package localoffsetmanager.config;
 
 import org.apache.kafka.clients.admin.NewTopic;
 import org.springframework.context.annotation.Bean;
@@ -11,7 +11,7 @@ public class TopicConfig {
   @Bean
   public NewTopic offsetTopic() {
     return TopicBuilder
-      .name("offset-topic")
+      .name("s3-offset-topic")
       .partitions(2)
       .replicas(3)
       .compact()
