+++ offset-manager-remote/src/main/java/localoffsetmanager/controller/dto/LastOffsetRecordBatchRequest.java
@@ -1,4 +1,4 @@
-package controller.dto;
+package localoffsetmanager.controller.dto;
 
 import jakarta.validation.constraints.NotEmpty;
