+++ offset-manager-remote/src/main/java/localoffsetmanager/error/ErrorMessage.java
@@ -1,0 +1,7 @@
+package localoffsetmanager.error;
+
+public record ErrorMessage(
+  String key,
+  String message
+) {
+}
