+++ offset-manager-remote/src/main/java/localoffsetmanager/service/dto/LastOffsetRecordResponse.java
@@ -1,4 +1,4 @@
-package service.dto;
+package localoffsetmanager.service.dto;
 
 import offsetmanager.domain.OffsetRecord;
